+++ tiled/mimetypes.py
@@ -19,6 +19,9 @@ DEFAULT_ADAPTERS_BY_MIMETYPE = OneShotCachedMap(
         "multipart/related;type=image/tiff": lambda: importlib.import_module(
             "..adapters.tiff", __name__
         ).TiffSequenceAdapter.from_uris,
+        "multipart/related;type=text/csv": lambda: importlib.import_module(
+            "..adapters.csv", __name__
+        ).CSVAdapter,
         "image/jpeg": lambda: importlib.import_module(
             "..adapters.jpeg", __name__
         ).JPEGAdapter,
@@ -28,9 +31,6 @@ DEFAULT_ADAPTERS_BY_MIMETYPE = OneShotCachedMap(
         "text/csv": lambda: importlib.import_module(
             "..adapters.csv", __name__
         ).CSVAdapter,
-        # "text/csv": lambda: importlib.import_module(
-        #    "..adapters.csv", __name__
-        # ).CSVAdapter.from_single_file,
         XLSX_MIME_TYPE: lambda: importlib.import_module(
             "..adapters.excel", __name__
         ).ExcelAdapter.from_uri,
@@ -62,9 +62,7 @@ DEFAULT_REGISTERATION_ADAPTERS_BY_MIMETYPE = copy.deepcopy(DEFAULT_ADAPTERS_BY_M
 
 DEFAULT_REGISTERATION_ADAPTERS_BY_MIMETYPE.set(
     "text/csv",
-    lambda: importlib.import_module(
-        "..adapters.csv", __name__
-    ).CSVAdapter.from_single_file,
+    lambda: importlib.import_module("..adapters.csv", __name__).CSVAdapter,
 )