+++ CHANGELOG.md
@@ -5,6 +5,11 @@ Write the date in place of the "Unreleased" in the case a new version is release
 
 ## Unreleased
 
+### Added
+
+- Automatic reshaping of tiff data by the adapter to account for
+  extra/missing singleton dimensions
+- Refactor CSVAdapter to allow pd.read_csv kwargs
 - Add adapters for reading back assets with the image/jpeg and
   multipart/related;type=image/jpeg mimetypes.