+++ CHANGELOG.md
@@ -3,7 +3,7 @@ Write the date in place of the "Unreleased" in the case a new version is release
 
 # Changelog
 
-## Unreleased
+## v0.1.0b11 (2024-11-14)
 
 ### Added